requirements.txt
name: seizuredetection-env
channels:
  - pytorch
  - conda-forge
  - defaults
dependencies:
  - _libgcc_mutex=0.1=conda_forge
  - _openmp_mutex=4.5=1_gnu
  - appdirs=1.4.4=pyh9f0ad1d_0
  - astroid=2.5.8=py38h06a4308_0
  - audioread=2.1.9=py38h578d9bd_0
  - blas=1.0=mkl
  - brotlipy=0.7.0=py38h497a2fe_1001
  - bzip2=1.0.8=h7b6447c_0
  - ca-certificates=2021.5.30=ha878542_0
  - cached-property=1.5.2=py_0
  - certifi=2021.5.30=py38h578d9bd_0
  - cffi=1.14.5=py38ha65f79e_0
  - chardet=4.0.0=py38h578d9bd_1
  - cpuonly=1.0=0
  - cryptography=3.4.7=py38ha5dfef3_0
  - cycler=0.10.0=py_2
  - decorator=5.0.9=pyhd8ed1ab_0
  - dill=0.3.3=pyhd3eb1b0_0
  - et_xmlfile=1.1.0=py38h06a4308_0
  - ffmpeg=4.3=hf484d3e_0
  - freetype=2.10.4=h5ab3b9f_0
  - gettext=0.19.8.1=h0b5b191_1005
  - gmp=6.2.1=h2531618_2
  - gnutls=3.6.15=he1e5248_0
  - h5py=3.2.1=py38h6c542dc_0
  - hdf5=1.10.6=hb1b8bf9_0
  - idna=2.10=pyh9f0ad1d_0
  - intel-openmp=2021.2.0=h06a4308_610
  - isort=5.8.0=pyhd3eb1b0_0
  - jdcal=1.4.1=py_0
  - joblib=1.0.1=pyhd8ed1ab_0
  - jpeg=9b=h024ee3a_2
  - kiwisolver=1.3.1=py38h1fd1430_1
  - lame=3.100=h7b6447c_0
  - lazy-object-proxy=1.6.0=py38h27cfd23_0
  - lcms2=2.12=h3be6417_0
  - ld_impl_linux-64=2.33.1=h53a641e_7
  - libffi=3.3=he6710b0_2
  - libflac=1.3.3=h9c3ff4c_1
  - libgcc-ng=9.3.0=h2828fa1_19
  - libgfortran-ng=7.5.0=h14aa051_19
  - libgfortran4=7.5.0=h14aa051_19
  - libgomp=9.3.0=h2828fa1_19
  - libiconv=1.15=h63c8f33_5
  - libidn2=2.3.1=h27cfd23_0
  - libllvm10=10.0.1=he513fc3_3
  - libogg=1.3.4=h7f98852_1
  - libopus=1.3.1=h7f98852_1
  - libpng=1.6.37=hbc83047_0
  - librosa=0.8.1=pyhd8ed1ab_0
  - libsndfile=1.0.31=h9c3ff4c_1
  - libstdcxx-ng=9.3.0=h6de172a_19
  - libtasn1=4.16.0=h27cfd23_0
  - libtiff=4.1.0=h2733197_1
  - libunistring=0.9.10=h27cfd23_0
  - libuv=1.40.0=h7b6447c_0
  - libvorbis=1.3.7=h9c3ff4c_0
  - lz4-c=1.9.3=h2531618_0
  - matplotlib-base=3.4.2=py38hcc49a3a_0
  - mccabe=0.6.1=py38_1
  - mkl=2021.2.0=h06a4308_296
  - mkl-service=2.3.0=py38h27cfd23_1
  - mkl_fft=1.3.0=py38h42c9631_2
  - mkl_random=1.2.1=py38ha9443f7_2
  - ncurses=6.2=he6710b0_1
  - nettle=3.7.2=hbbd107a_1
  - ninja=1.10.2=hff7bd54_1
  - numba=0.53.1=py38ha9443f7_0
  - numpy=1.20.2=py38h2d18471_0
  - numpy-base=1.20.2=py38hfae3a4d_0
  - olefile=0.46=py_0
  - openh264=2.1.0=hd408876_0
  - openpyxl=3.0.7=pyhd3eb1b0_0
  - openssl=1.1.1k=h7f98852_0
  - packaging=20.9=pyh44b312d_0
  - pillow=8.2.0=py38he98fc37_0
  - pip=21.1.1=py38h06a4308_0
  - pooch=1.3.0=pyhd8ed1ab_0
  - progressbar=2.5=py_0
  - pycparser=2.20=pyh9f0ad1d_2
  - pyedflib=0.1.22=py38hb5d20a5_0
  - pylint=2.8.3=py38h06a4308_1
  - pyopenssl=20.0.1=pyhd8ed1ab_0
  - pyparsing=2.4.7=pyh9f0ad1d_0
  - pysocks=1.7.1=py38h578d9bd_3
  - pysoundfile=0.10.3.post1=pyhd3deb0d_0
  - python=3.8.10=hdb3f193_7
  - python-dateutil=2.8.1=py_0
  - python-utils=2.5.6=pyh44b312d_0
  - python_abi=3.8=1_cp38
  - pytorch=1.8.1=py3.8_cpu_0
  - pywavelets=1.1.1=py38h5c078b8_3
  - readline=8.1=h27cfd23_0
  - regex=2021.4.4=py38h497a2fe_0
  - requests=2.25.1=pyhd3deb0d_0
  - resampy=0.2.2=py_0
  - scikit-learn=0.24.2=py38ha9443f7_0
  - setuptools=52.0.0=py38h06a4308_0
  - six=1.15.0=py38h06a4308_0
  - sqlite=3.35.4=hdfb4753_0
  - tabulate=0.8.9=py38h06a4308_0
  - tbb=2020.2=hc9558a2_0
  - threadpoolctl=2.1.0=pyh5ca1d4c_0
  - tk=8.6.10=hbc83047_0
  - toml=0.10.2=pyhd3eb1b0_0
  - torchaudio=0.8.1=py38
  - torchvision=0.9.1=py38_cpu
  - tornado=6.1=py38h497a2fe_1
  - typing_extensions=3.7.4.3=pyha847dfd_0
  - urllib3=1.26.5=pyhd8ed1ab_0
  - wheel=0.36.2=pyhd3eb1b0_0
  - wrapt=1.12.1=py38h7b6447c_1
  - xlrd=2.0.1=pyhd3eb1b0_0
  - xz=5.2.5=h7b6447c_0
  - zlib=1.2.11=h7b6447c_3
  - zstd=1.4.9=haebb681_0
  - pip:
    - antropy==0.1.4
    - argon2-cffi==20.1.0
    - async-generator==1.10
    - attrs==21.2.0
    - backcall==0.2.0
    - bleach==3.3.0
    - dataclasses==0.6
    - defusedxml==0.7.1
    - entrypoints==0.3
    - greenlet==1.1.0
    - ipykernel==5.5.5
    - ipython==7.24.1
    - ipython-genutils==0.2.0
    - ipywidgets==7.6.3
    - jedi==0.18.0
    - jinja2==3.0.1
    - jsonschema==3.2.0
    - jupyter-client==6.1.12
    - jupyter-core==4.7.1
    - jupyterlab-pygments==0.1.2
    - jupyterlab-widgets==1.0.0
    - littleutils==0.2.2
    - llvmlite==0.36.0
    - lspopt==1.1.1
    - markupsafe==2.0.1
    - matplotlib-inline==0.1.2
    - mistune==0.8.4
    - mne==0.23.0
    - nbclient==0.5.3
    - nbconvert==6.0.7
    - nbformat==5.1.3
    - nest-asyncio==1.5.1
    - notebook==6.4.0
    - outdated==0.2.1
    - pandas==1.2.4
    - pandocfilters==1.4.3
    - parso==0.8.2
    - pexpect==4.8.0
    - pickleshare==0.7.5
    - prometheus-client==0.11.0
    - prompt-toolkit==3.0.18
    - ptyprocess==0.7.0
    - pyeeglab==0.10.0
    - pygments==2.9.0
    - pyriemann==0.2.6
    - pyrsistent==0.17.3
    - pytz==2021.1
    - pyzmq==22.1.0
    - scipy==1.6.3
    - seaborn==0.11.1
    - send2trash==1.5.0
    - sqlalchemy==1.4.17
    - stft==0.5.2
    - stochastic==0.6.0
    - tensorpac==0.6.5
    - terminado==0.10.0
    - testpath==0.5.0
    - traitlets==5.0.5
    - wcwidth==0.2.5
    - webencodings==0.5.1
    - wfdb==3.4.0
    - widgetsnbextension==3.5.1
    - yasa==0.4.1
prefix: /home/dominika/anaconda3/envs/seizuredetection-env

numpy~=1.20.2
scipy~=1.6.3
scikit-learn~=0.24.2
librosa~=0.8.1
pandas~=1.2.4
pyedflib~=0.1.22
progressbar~=2.5
tabulate~=0.8.9
dill~=0.3.3
utils~=1.0.1
resampy~=0.2.2
regex~=2021.4.4
matplotlib~=3.4.2
seaborn~=0.11.1
wandb~=0.11.0
torchvision~=0.9.1
pyeeglab~=0.10.2
torchsampler~=0.1.1
joblib~=1.0.1